services/monitor/requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
plotly
requests
pyyaml
